Move this method into utils because it should be not use in other place.
