+++ osu.Game.Rulesets.Karaoke.Tests/Beatmaps/Formats/KaraokeLegacyBeatmapDecoderTest.cs
@@ -10,6 +10,7 @@ using osu.Game.IO;
 using osu.Game.Rulesets.Karaoke.Beatmaps;
 using osu.Game.Rulesets.Karaoke.Beatmaps.Formats;
 using osu.Game.Rulesets.Karaoke.Objects;
+using osu.Game.Rulesets.Karaoke.Tests.Helper;
 using osu.Game.Rulesets.Karaoke.Tests.Resources;
 using osu.Game.Rulesets.Mods;
 using osu.Game.Tests.Beatmaps;
@@ -128,6 +129,38 @@ namespace osu.Game.Rulesets.Karaoke.Tests.Beatmaps.Formats
             Assert.AreEqual(expectedText, actualNote.Text);
             Assert.AreEqual(expectedTone, actualNote.Tone.Scale);
             Assert.AreEqual(expectedHalf, actualNote.Tone.Half);
+        }
+
+        [TestCase(0, 1, new double[] { 1000, 3000 })]
+        [TestCase(0, 0.5, new double[] { 1000, 1500 })]
+        [TestCase(0.5, 0.5, new double[] { 2500, 1500 })]
+        [TestCase(0.3, 0.4, new double[] { 1900, 1200 })]
+        [TestCase(0.3, 1, null)] // start + duration should not exceed 1
+        public void TestSliceNoteTime(double startPercentage, double durationPercentage, double[]? expected)
+        {
+            var lyric = new Lyric
+            {
+                TimeTags = TestCaseTagHelper.ParseTimeTags(new[] { "[0,start]:1000", "[1,start]:4000" }),
+            };
+
+            // start time will be 1000, and duration will be 3000.
+            var note = new Note
+            {
+                ReferenceLyric = lyric,
+                ReferenceTimeTagIndex = 0
+            };
+
+            if (expected != null)
+            {
+                var sliceNote = KaraokeLegacyBeatmapDecoder.SliceNote(note, startPercentage, durationPercentage);
+
+                Assert.AreEqual(expected[0], sliceNote.StartTime);
+                Assert.AreEqual(expected[1], sliceNote.Duration);
+            }
+            else
+            {
+                Assert.IsNull(expected);
+            }
         }
     }
 }
